Sparc: fix cache coherency issue
Problem:
As the cache is virtually-tagged, we can have incoherency between
multiple mappings of the same virtual address. In practice, the when
the user writes to the utcb, the kernel may still read the old value
from the data cache since it uses its own mapping for the user's utcb.

Solution:
Flushing the data cache on syscall entry & exit.